fix: preserve iOS presentation context for scoped snapshots
