+++ src/daemon/snapshot-presentation/ios/action-shelf.ts
@@ -1,0 +1,248 @@
+import type { RawSnapshotNode, Rect } from '@agent-device/kernel/snapshot';
+import { isRectVisibleInViewport, rectContains } from '@agent-device/kernel/rect';
+import { normalizeType } from '@agent-device/contracts/snapshot';
+import {
+  areRectsApproximatelyEqual,
+  collectDescendants,
+  findDescendant,
+  isSemanticActionNode,
+  isScrollableSnapshotType,
+  type SnapshotTreeRuleContext,
+} from '../tree.ts';
+
+const ACTION_SHELF_MINIMUM_BUTTONS = 3;
+const ACTION_SHELF_EDGE_TOLERANCE = 2;
+// A rotating add/close toggle grows its axis-aligned AX bounds toward sqrt(2). Keep the threshold
+// well below that transform while leaving ordinary pixel jitter and same-sized shelf actions alone.
+const ACTION_TOGGLE_EXPANSION_RATIO = 1.15;
+
+interface ActionShelfPresentation {
+  rect: Rect;
+  childActionsPresented: boolean;
+  actionButtons: RawSnapshotNode[];
+}
+
+export function collectIosReplacedActionShelves(
+  nodes: RawSnapshotNode[],
+  childrenByParent: ReadonlyMap<number, RawSnapshotNode[]>,
+  context: SnapshotTreeRuleContext,
+): void {
+  const positions = new Map(nodes.map((node, position) => [node.index, position]));
+  for (const [shelfPosition, shelf] of nodes.entries()) {
+    const transition = resolveReplacedActionShelf(
+      nodes,
+      positions,
+      childrenByParent,
+      shelf,
+      context,
+    );
+    if (!transition) continue;
+
+    suppressNodeAndDescendants(nodes, shelfPosition, context.suppressedIndexes);
+    const shelfBand = expandRect(transition.shelfRect, ACTION_SHELF_EDGE_TOLERANCE);
+    for (const sibling of transition.intermediateSiblings) {
+      if (
+        isTextInput(sibling) ||
+        !sibling.rect ||
+        !isRectVisibleInViewport(sibling.rect, shelfBand)
+      ) {
+        continue;
+      }
+      const position = positions.get(sibling.index);
+      if (position !== undefined) {
+        suppressNodeAndDescendants(nodes, position, context.suppressedIndexes);
+      }
+    }
+  }
+}
+
+function resolveReplacedActionShelf(
+  nodes: RawSnapshotNode[],
+  positions: ReadonlyMap<number, number>,
+  childrenByParent: ReadonlyMap<number, RawSnapshotNode[]>,
+  shelf: RawSnapshotNode,
+  context: SnapshotTreeRuleContext,
+): { shelfRect: Rect; intermediateSiblings: RawSnapshotNode[] } | undefined {
+  const hiddenShelf = resolveHiddenActionShelf(shelf, childrenByParent, context.sourceNodesByIndex);
+  if (!hiddenShelf) return undefined;
+  const replacementPosition = findActionShelfReplacementPosition(
+    nodes,
+    positions,
+    hiddenShelf.siblings,
+    hiddenShelf.position,
+    hiddenShelf.parent,
+  );
+  if (replacementPosition < 0) return undefined;
+  return {
+    shelfRect: hiddenShelf.rect,
+    intermediateSiblings: hiddenShelf.siblings.slice(hiddenShelf.position + 1, replacementPosition),
+  };
+}
+
+function resolveHiddenActionShelf(
+  shelf: RawSnapshotNode,
+  childrenByParent: ReadonlyMap<number, RawSnapshotNode[]>,
+  nodesByIndex: ReadonlyMap<number, RawSnapshotNode>,
+):
+  | { rect: Rect; parent: RawSnapshotNode; siblings: RawSnapshotNode[]; position: number }
+  | undefined {
+  const presentation = resolveHorizontalActionShelf(
+    shelf,
+    childrenFor(childrenByParent, shelf.index),
+  );
+  if (!presentation) return undefined;
+  const parent = findParent(shelf, nodesByIndex);
+  if (!parent || !parent.rect) return undefined;
+  const siblings = childrenFor(childrenByParent, parent.index);
+  const position = siblings.findIndex((node) => node.index === shelf.index);
+  if (!hasCollapsedShelfToggle(siblings, position, shelf, presentation)) return undefined;
+  return { rect: presentation.rect, parent, siblings, position };
+}
+
+function childrenFor(
+  childrenByParent: ReadonlyMap<number, RawSnapshotNode[]>,
+  parentIndex: number,
+): RawSnapshotNode[] {
+  return childrenByParent.get(parentIndex) ?? [];
+}
+
+function hasCollapsedShelfToggle(
+  siblings: RawSnapshotNode[],
+  shelfPosition: number,
+  shelf: RawSnapshotNode,
+  presentation: ActionShelfPresentation,
+): boolean {
+  if (shelfPosition < 1) return false;
+  const leadingAction = findIndependentLeadingAction(siblings, shelfPosition, shelf);
+  if (!leadingAction) return false;
+  if (!presentation.childActionsPresented) return true;
+  return !isExpandedActionToggle(leadingAction, presentation.actionButtons);
+}
+
+function findParent(
+  node: RawSnapshotNode,
+  nodesByIndex: ReadonlyMap<number, RawSnapshotNode>,
+): RawSnapshotNode | undefined {
+  return typeof node.parentIndex === 'number' ? nodesByIndex.get(node.parentIndex) : undefined;
+}
+
+function findActionShelfReplacementPosition(
+  nodes: RawSnapshotNode[],
+  positions: ReadonlyMap<number, number>,
+  siblings: RawSnapshotNode[],
+  shelfPosition: number,
+  parent: RawSnapshotNode,
+): number {
+  return siblings.findIndex(
+    (candidate, position) =>
+      position > shelfPosition &&
+      isActionShelfReplacement(nodes, positions.get(candidate.index), candidate, parent),
+  );
+}
+
+function resolveHorizontalActionShelf(
+  shelf: RawSnapshotNode,
+  children: RawSnapshotNode[],
+): ActionShelfPresentation | undefined {
+  if (!isScrollableSnapshotType(shelf.type) || !shelf.rect) return undefined;
+  const shelfRect = shelf.rect;
+  const buttons = children.filter(
+    (node) => normalizeType(node.type ?? '') === 'button' && node.rect,
+  );
+  if (buttons.length === 0 && shelf.hiddenContentBelow === true) {
+    return { rect: shelfRect, childActionsPresented: false, actionButtons: buttons };
+  }
+  if (buttons.length < ACTION_SHELF_MINIMUM_BUTTONS) return undefined;
+  const centers = buttons.map((button) => button.rect!.x + button.rect!.width / 2);
+  if (!centers.every((center, index) => index === 0 || center > centers[index - 1]!)) {
+    return undefined;
+  }
+  return {
+    rect: shelfRect,
+    childActionsPresented: buttons.some((button) => rectContains(shelfRect, button.rect!)),
+    actionButtons: buttons,
+  };
+}
+
+function findIndependentLeadingAction(
+  siblings: RawSnapshotNode[],
+  shelfPosition: number,
+  shelf: RawSnapshotNode,
+): RawSnapshotNode | undefined {
+  if (!shelf.rect || shelfPosition < 1) return undefined;
+  for (let position = shelfPosition - 1; position >= 0; position -= 1) {
+    const candidate = siblings[position]!;
+    if (!candidate.rect || !isSemanticActionNode(candidate)) continue;
+    const centerX = candidate.rect.x + candidate.rect.width / 2;
+    if (
+      centerX < shelf.rect.x &&
+      verticallyOverlaps(candidate.rect, expandRect(shelf.rect, ACTION_SHELF_EDGE_TOLERANCE))
+    ) {
+      return candidate;
+    }
+  }
+  return undefined;
+}
+
+function isExpandedActionToggle(
+  toggle: RawSnapshotNode,
+  actionButtons: RawSnapshotNode[],
+): boolean {
+  if (!toggle.rect || actionButtons.length === 0) return false;
+  const actionWidth = Math.max(...actionButtons.map((button) => button.rect?.width ?? 0));
+  const actionHeight = Math.max(...actionButtons.map((button) => button.rect?.height ?? 0));
+  return (
+    toggle.rect.width > actionWidth * ACTION_TOGGLE_EXPANSION_RATIO ||
+    toggle.rect.height > actionHeight * ACTION_TOGGLE_EXPANSION_RATIO
+  );
+}
+
+function isActionShelfReplacement(
+  nodes: RawSnapshotNode[],
+  position: number | undefined,
+  candidate: RawSnapshotNode,
+  parent: RawSnapshotNode,
+): boolean {
+  return (
+    position !== undefined &&
+    normalizeType(candidate.type ?? '') === 'other' &&
+    areRectsApproximatelyEqual(candidate.rect, parent.rect) &&
+    Boolean(findDescendant(nodes, position, isSemanticActionNode))
+  );
+}
+
+function suppressNodeAndDescendants(
+  nodes: RawSnapshotNode[],
+  position: number,
+  suppressedIndexes: Set<number>,
+): void {
+  const node = nodes[position];
+  if (!node) return;
+  suppressedIndexes.add(node.index);
+  for (const descendant of collectDescendants(nodes, position)) {
+    suppressedIndexes.add(descendant.index);
+  }
+}
+
+function isTextInput(node: RawSnapshotNode): boolean {
+  const type = normalizeType(node.type ?? '');
+  return (
+    type === 'textfield' ||
+    type === 'securetextfield' ||
+    type === 'searchfield' ||
+    type === 'textview'
+  );
+}
+
+function expandRect(rect: Rect, amount: number): Rect {
+  return {
+    x: rect.x - amount,
+    y: rect.y - amount,
+    width: rect.width + amount * 2,
+    height: rect.height + amount * 2,
+  };
+}
+
+function verticallyOverlaps(left: Rect, right: Rect): boolean {
+  return Math.max(left.y, right.y) <= Math.min(left.y + left.height, right.y + right.height);
+}
